# More reduced allocations