--- a/test/cases/testdata/v1/walkbuiltin/test-walkbuiltin-0971.yaml
+++ b/test/cases/testdata/v1/walkbuiltin/test-walkbuiltin-0971.yaml
@@ -32,3 +32,27 @@ cases:
           - - - 3
             - 4
     sort_bindings: true
+  - note: walkbuiltin/arrays no path
+    query: data.test.p = x
+    modules:
+      - |
+        package test
+
+        p := [x | walk(data.a, [_, x])]
+    data:
+      a:
+        - 1
+        - 2
+        - 3
+        - 4
+    want_result:
+      - x:
+          - 1
+          - 2
+          - 3
+          - 4
+          - - 1
+            - 2
+            - 3
+            - 4
+    sort_bindings: true