--- a/test/cases/testdata/v1/walkbuiltin/test-walkbuiltin-0972.yaml
+++ b/test/cases/testdata/v1/walkbuiltin/test-walkbuiltin-0972.yaml
@@ -24,3 +24,21 @@ cases:
           - - - v2
             - goodbye
     sort_bindings: true
+  - note: walkbuiltin/objects no path
+    query: data.generated.p = x
+    modules:
+      - |
+        package generated
+
+        p := [x | walk(data.b, [_, x])]
+    data:
+      b:
+        v1: hello
+        v2: goodbye
+    want_result:
+      - x:
+          - goodbye
+          - hello
+          - v1: hello
+            v2: goodbye
+    sort_bindings: true