@anderseknert 2024-11-27
The failing test is "unrelated to this PR", which is a bit of a bold statement, given that the test was added by this very PR. But that test fails even on current OPA, and it seems like we have some ordering issue in the Wasm-implementation of `walk`.

```
 --- FAIL: TestWasmE2E (308.07s)
    --- FAIL: TestWasmE2E/test/cases/testdata/v1/walkbuiltin/test-walkbuiltin-0972.yaml/walkbuiltin/objects_no_path (0.09s)
        external_test.go:110: expected {{"x": [{"v1": "hello", "v2": "goodbye"}, "hello", "goodbye"]}} but got {{"x": [{"v1": "hello", "v2": "goodbye"}, "goodbye", "hello"]}}
```